Feat: 소수점 자리수 정리 (#26) - BigDecimal 전체 변경

+++ src/main/java/com/project/easywork/common/util/measurePoint/MeasurePointStrategy.java
@@ -1,5 +1,7 @@
 package com.project.easywork.common.util.measurePoint;
 
+import java.math.BigDecimal;
+
 public interface MeasurePointStrategy {
-  Integer calculate(double...values);
+  Integer calculate(BigDecimal...values);
 }
